requirements.txt
aiohttp~=3.12.13
customtkinter~=5.2.2
opencv-python~=4.11.0.86
numpy~=2.3.0
onnxruntime-directml
pyautogui~=0.9.54
pygetwindow~=0.0.9
requests~=2.32.4
toml~=0.10.2
pillow~=11.2.1
discord.py
shapely~=2.1.1
packaging~=25.0
pywin32
pure-python-adb
google-play-scraper~=1.2.7
easyocr~=1.7.2
av~=12.3.0
bettercam
